Move get_country_code.py and countries.py into ./src/django/countries/lib/
