+++ src/django/api/models/facility/facility_list_item.py
@@ -1,4 +1,4 @@
-from countries.countries import COUNTRY_CHOICES
+from countries.lib.countries import COUNTRY_CHOICES
 
 from django.contrib.gis.db import models as gis_models
 from django.contrib.postgres import fields as postgres
